Added functionality for Android "panel indicator"
Added an additional optional property called "indicatorId" which takes the ID of a DOM element to be used as a sidebar indicator such as those provided by the native holo interface in Android. The chosen indicator is slid in and out in proportion to the sidebar position just as in the native interface element.